Added assinmentscores to the lower/upperbound

+++ Qualification/Qualification/BrandAndBound/Assignment.cs
@@ -2,7 +2,7 @@
 {
     public class Assignment
     {
-        public double Value { get; set; }
+        public int Value { get; set; }
         public int VehicleId { get; set; }
         public int RideId { get; set; }
     }
